+++ lib/puppet-debugger.rb
@@ -1,12 +1,12 @@
-require_relative 'puppet-repl/cli'
+require_relative 'puppet-debugger/cli'
 require_relative 'version'
 require 'awesome_print'
 require_relative 'awesome_print/ext/awesome_puppet'
 require_relative 'trollop'
 require 'puppet/util/log'
-require_relative 'puppet-repl/repl_code'
+require_relative 'puppet-debugger/debugger_code'
 
-require_relative 'puppet-repl/support/errors'
+require_relative 'puppet-debugger/support/errors'
 # monkey patch in some color effects string methods
 class String
   def red;            "\033[31m#{self}\033[0m" end