+++ lib/puppet-debugger/code/code_range.rb
@@ -1,4 +1,4 @@
-  class ReplCode
+  class DebuggerCode
 
     # Represents a range of lines in a code listing.
     #