+++ lib/puppet-debugger/code/loc.rb
@@ -1,4 +1,4 @@
-  class ReplCode
+  class DebuggerCode
 
     # Represents a line of code. A line of code is a tuple, which consists of a
     # line and a line number. A `LOC` object's state (namely, the line