+++ spec/puppet-debugger_spec.rb
@@ -1,21 +1,21 @@
 require 'spec_helper'
 require 'stringio'
-describe "PuppetRepl" do
+describe "PuppetDebugger" do
 
   let(:resource) do
     "service{'httpd': ensure => running}"
   end
 
   before(:each) do
-    repl.handle_input('reset')
+    debugger.handle_input('reset')
   end
 
   let(:output) do
     StringIO.new('', 'w')
   end
 
-  let(:repl) do
-    PuppetRepl::Cli.new({:out_buffer => output}.merge(options))
+  let(:debugger) do
+    PuppetDebugger::Cli.new({:out_buffer => output}.merge(options))
   end
 
   let(:options) do
@@ -27,7 +27,7 @@ describe "PuppetRepl" do
   end
 
   let(:resource_types) do
-    repl.parser.evaluate_string(repl.scope, input)
+    debugger.parser.evaluate_string(debugger.scope, input)
   end
 
   describe 'native classes' do
@@ -35,23 +35,23 @@ describe "PuppetRepl" do
       let(:input) do
         'class testfoo {}'
       end
-      let(:repl_output) do
+      let(:debugger_output) do
         "\n => Puppet::Type::Component {\n  loglevel\e[0;37m => \e[0m\e[0;36mnotice\e[0m,\n      name\e[0;37m => \e[0m\e[0;33m\"Testfoo\"\e[0m,\n     title\e[0;37m => \e[0m\e[0;33m\"Class[Testfoo]\"\e[0m\n}\n"
       end
       it do
-        repl.handle_input(input)
+        debugger.handle_input(input)
         expect(output.string).to eq("\n")
-        expect(repl.known_resource_types[:hostclasses]).to include('testfoo')
+        expect(debugger.known_resource_types[:hostclasses]).to include('testfoo')
       end
       it do
-        repl.handle_input(input)
-        repl.handle_input("include testfoo")
-        expect(repl.scope.compiler.catalog.classes).to include('testfoo')
+        debugger.handle_input(input)
+        debugger.handle_input("include testfoo")
+        expect(debugger.scope.compiler.catalog.classes).to include('testfoo')
       end
       it do
-        repl.handle_input(input)
-        repl.handle_input('include testfoo')
-        expect(repl.scope.compiler.catalog.resources.map(&:name)).to include('Testfoo')
+        debugger.handle_input(input)
+        debugger.handle_input('include testfoo')
+        expect(debugger.scope.compiler.catalog.resources.map(&:name)).to include('Testfoo')
       end
     end
   end
@@ -61,19 +61,19 @@ describe "PuppetRepl" do
       let(:input) do
         'define testfoo {}'
       end
-      let(:repl_output) do
+      let(:debugger_output) do
         "\n => Puppet::Type::Component {\n  loglevel\e[0;37m => \e[0m\e[0;36mnotice\e[0m,\n      name\e[0;37m => \e[0m\e[0;33m\"some_name\"\e[0m,\n     title\e[0;37m => \e[0m\e[0;33m\"Testfoo[some_name]\"\e[0m\n}\n"
       end
       it do
-        repl.handle_input(input)
-        expect(repl.scope.environment.known_resource_types.definitions.keys).to include('testfoo')
+        debugger.handle_input(input)
+        expect(debugger.scope.environment.known_resource_types.definitions.keys).to include('testfoo')
         expect(output.string).to eq("\n")
       end
       it do
-        repl.handle_input(input)
-        repl.handle_input("testfoo{'some_name':}")
-        expect(repl.scope.compiler.resources.collect(&:name)).to include('some_name')
-        expect(repl.scope.compiler.resources.collect(&:type)).to include('Testfoo')
+        debugger.handle_input(input)
+        debugger.handle_input("testfoo{'some_name':}")
+        expect(debugger.scope.compiler.resources.collect(&:name)).to include('some_name')
+        expect(debugger.scope.compiler.resources.collect(&:type)).to include('Testfoo')
         expect(output.string).to include("\n => Puppet::Type::Component")
       end
     end
@@ -82,7 +82,7 @@ describe "PuppetRepl" do
   describe 'native functions', :native_functions => true do
     let(:func) do
       <<-EOF
-      function repl::bool2http($arg) {
+      function debugger::bool2http($arg) {
         case $arg {
           false, undef, /(?i:false)/ : { 'Off' }
           true, /(?i:true)/          : { 'On' }
@@ -92,7 +92,7 @@ describe "PuppetRepl" do
       EOF
     end
     before(:each) do
-      repl.handle_input(func)
+      debugger.handle_input(func)
     end
     describe 'create' do
       it 'shows function' do
@@ -102,11 +102,11 @@ describe "PuppetRepl" do
     describe 'run' do
       let(:input) do
         <<-EOF
-        repl::bool2http(false)
+        debugger::bool2http(false)
         EOF
       end
       it do
-        repl.handle_input(input)
+        debugger.handle_input(input)
         expect(output.string).to include('Off')
       end
     end
@@ -119,7 +119,7 @@ describe "PuppetRepl" do
         "String"
       end
       it 'shows type' do
-        repl.handle_input(input)
+        debugger.handle_input(input)
         expect(output.string).to eq("\n => String\n")
       end
     end
@@ -128,7 +128,7 @@ describe "PuppetRepl" do
         "type_of([1,2,3,4])"
       end
       it 'shows type' do
-        repl.handle_input(input)
+        debugger.handle_input(input)
         expect(output.string).to eq("\n => Tuple[Integer[1, 1], Integer[2, 2], Integer[3, 3], Integer[4, 4]]\n")
       end
     end
@@ -141,7 +141,7 @@ describe "PuppetRepl" do
         "$var1 = 'test'\nfile{\"/tmp/${var1}.txt\": ensure => present, mode => '0755'}\nvars"
       end
       it do
-        repl.play_back_string(input)
+        debugger.play_back_string(input)
         expect(output.string).to match(/server_facts/) if Puppet.version.to_f >= 4.1
         expect(output.string).to match(/test/)
         expect(output.string).to match(/Puppet::Type::File/)
@@ -152,7 +152,7 @@ describe "PuppetRepl" do
         "$var1 = 'test'\n $var2 = 'test2'"
       end
       it do
-        repl.play_back_string(input)
+        debugger.play_back_string(input)
         expect(output.string).to include("$var1 = 'test'")
         expect(output.string).to include("\"test\"")
         expect(output.string).to include("$var2 = 'test2'")
@@ -164,7 +164,7 @@ describe "PuppetRepl" do
         "$var1 = 'test'"
       end
       it do
-        repl.play_back_string(input)
+        debugger.play_back_string(input)
         expect(output.string).to include("$var1 = 'test'")
         expect(output.string).to include("\"test\"")
       end
@@ -182,13 +182,13 @@ describe "PuppetRepl" do
       'help'
     end
     it 'can show the help screen' do
-      expected_repl_output = /Type \"exit\", \"functions\", \"vars\", \"krt\", \"whereami\", \"facts\", \"resources\", \"classes\",\n     \"play\", \"classification\", \"reset\", or \"help\" for more information.\n\n/
-      repl.handle_input(input)
+      expected_debugger_output = /Type \"exit\", \"functions\", \"vars\", \"krt\", \"whereami\", \"facts\", \"resources\", \"classes\",\n     \"play\", \"classification\", \"reset\", or \"help\" for more information.\n\n/
+      debugger.handle_input(input)
       expect(output.string).to match(/Ruby Version: #{RUBY_VERSION}\n/)
       expect(output.string).to match(/Puppet Version: \d.\d.\d\n/)
-      expect(output.string).to match(/Puppet Repl Version: \d.\d.\d\n/)
+      expect(output.string).to match(/Puppet Debugger Version: \d.\d.\d\n/)
       expect(output.string).to match(/Created by: NWOps <corey@nwops.io>\n/)
-      expect(output.string).to match(expected_repl_output)
+      expect(output.string).to match(expected_debugger_output)
     end
   end
 
@@ -197,18 +197,18 @@ describe "PuppetRepl" do
       ""
     end
     it 'can run' do
-      repl_output = "\n"
-      repl.handle_input(input)
-      expect(output.string).to eq(repl_output)
+      debugger_output = "\n"
+      debugger.handle_input(input)
+      expect(output.string).to eq(debugger_output)
     end
     describe 'space' do
       let(:input) do
         " "
       end
       it 'can run' do
-        repl_output = "\n"
-        repl.handle_input(input)
-        expect(output.string).to eq(repl_output)
+        debugger_output = "\n"
+        debugger.handle_input(input)
+        expect(output.string).to eq(debugger_output)
       end
     end
   end
@@ -218,9 +218,9 @@ describe "PuppetRepl" do
       "krt"
     end
     it 'can run' do
-      repl_output = /hostclasses/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /hostclasses/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -230,18 +230,18 @@ describe "PuppetRepl" do
     end
 
     before(:each) do
-      allow(repl).to receive(:fetch_url_data).with(file_url + '.txt').and_return(File.read(fixtures_file))
+      allow(debugger).to receive(:fetch_url_data).with(file_url + '.txt').and_return(File.read(fixtures_file))
     end
 
     let(:file_url) do
       'https://gist.githubusercontent.com/logicminds/f9b1ac65a3a440d562b0'
     end
     it 'file' do
-      repl.handle_input("play #{fixtures_file}")
+      debugger.handle_input("play #{fixtures_file}")
       expect(output.string).to match(/Puppet::Type::File/)
     end
     it 'url' do
-      repl.handle_input("play #{file_url}")
+      debugger.handle_input("play #{file_url}")
       expect(output.string).to match(/Puppet::Type::File/)
     end
   end
@@ -251,7 +251,7 @@ describe "PuppetRepl" do
       "$file_path = '/tmp/test2.txt'"
     end
     it 'can process a variable' do
-      repl.handle_input(input)
+      debugger.handle_input(input)
       expect(output.string).to match(/\/tmp\/test2.txt/)
     end
   end
@@ -261,9 +261,9 @@ describe "PuppetRepl" do
       "file{'/tmp/test2.txt': ensure => present, mode => '0755'}"
     end
     it 'can process a resource' do
-      repl_output = /Puppet::Type::File/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /Puppet::Type::File/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -272,9 +272,9 @@ describe "PuppetRepl" do
       "Service{"
     end
     it 'can process' do
-      repl_output = "\n => \e[31mSyntax error at end of file\e[0m\n"
-      repl.handle_input(input)
-      expect(output.string).to eq(repl_output)
+      debugger_output = "\n => \e[31mSyntax error at end of file\e[0m\n"
+      debugger.handle_input(input)
+      expect(output.string).to eq(debugger_output)
     end
   end
 
@@ -284,7 +284,7 @@ describe "PuppetRepl" do
     end
 
     it 'can process a file' do
-      repl.handle_input(input)
+      debugger.handle_input(input)
       expect(output.string).to eq("\n[]\n")
     end
   end
@@ -295,20 +295,20 @@ describe "PuppetRepl" do
     end
 
     it 'can process a file' do
-      repl_output = /Puppet::Type::File/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
-      repl.handle_input('reset')
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /Puppet::Type::File/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
+      debugger.handle_input('reset')
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
 
     describe 'loglevel' do
       it 'has not changed' do
-        repl.handle_input(":set loglevel debug")
+        debugger.handle_input(":set loglevel debug")
         expect(Puppet::Util::Log.level).to eq(:debug)
         expect(Puppet::Util::Log.destinations[:buffer].name).to eq(:buffer)
-        repl.handle_input('reset')
+        debugger.handle_input('reset')
         expect(Puppet::Util::Log.level).to eq(:debug)
         expect(Puppet::Util::Log.destinations[:buffer].name).to eq(:buffer)
       end
@@ -320,9 +320,9 @@ describe "PuppetRepl" do
       "['/tmp/test3', '/tmp/test4'].map |String $path| { file{$path: ensure => present} }"
     end
     it 'can process a each block' do
-      repl_output = /Puppet::Type::File/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /Puppet::Type::File/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -331,7 +331,7 @@ describe "PuppetRepl" do
       "['/tmp/test3', '/tmp/test4'].each |String $path| { file{$path: ensure => present} }"
     end
     it 'can process a each block' do
-      repl.handle_input(input)
+      debugger.handle_input(input)
       expect(output.string).to match(/\/tmp\/test3/)
       expect(output.string).to match(/\/tmp\/test4/)
     end
@@ -342,9 +342,9 @@ describe "PuppetRepl" do
       "$::fqdn"
     end
     it 'should be able to resolve fqdn' do
-      repl_output = /foo\.example\.com/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /foo\.example\.com/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -353,9 +353,9 @@ describe "PuppetRepl" do
       "facts"
     end
     it 'should be able to print facts' do
-      repl_output = /kernel/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /kernel/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -364,9 +364,9 @@ describe "PuppetRepl" do
       'resources'
     end
     it 'should be able to print resources' do
-      repl_output = /main/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /main/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -375,9 +375,9 @@ describe "PuppetRepl" do
       "Class['settings']"
     end
     it 'should be able to print classes' do
-      repl_output = /Settings/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /Settings/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -386,9 +386,9 @@ describe "PuppetRepl" do
       'resources'
     end
     it 'should be able to print classes' do
-      repl_output = /Settings/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /Settings/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
   end
 
@@ -397,9 +397,9 @@ describe "PuppetRepl" do
       ":set loglevel debug"
     end
     it 'should set the loglevel' do
-      repl_output = /loglevel debug is set/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /loglevel debug is set/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
       expect(Puppet::Util::Log.level).to eq(:debug)
       expect(Puppet::Util::Log.destinations[:buffer].name).to eq(:buffer)
     end
@@ -410,24 +410,24 @@ describe "PuppetRepl" do
       "vars"
     end
     it 'display facts variable' do
-      repl_output = /facts/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output = /facts/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
     it 'display server facts variable' do
-      repl_output = /server_facts/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output) if Puppet.version.to_f >= 4.1
+      debugger_output = /server_facts/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output) if Puppet.version.to_f >= 4.1
     end
     it 'display serverversion variable' do
-      repl_output = /serverversion/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output) if Puppet.version.to_f >= 4.1
+      debugger_output = /serverversion/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output) if Puppet.version.to_f >= 4.1
     end
     it 'display local variable' do
-      repl.handle_input("$var1 = 'value1'")
+      debugger.handle_input("$var1 = 'value1'")
       expect(output.string).to match(/value1/)
-      repl.handle_input("$var1")
+      debugger.handle_input("$var1")
       expect(output.string).to match(/value1/)
     end
 
@@ -438,20 +438,20 @@ describe "PuppetRepl" do
       "md5('hello')"
     end
     it 'execute md5' do
-      repl_output =  /5d41402abc4b2a76b9719d911017c592/
-      repl.handle_input(input)
-      expect(output.string).to match(repl_output)
+      debugger_output =  /5d41402abc4b2a76b9719d911017c592/
+      debugger.handle_input(input)
+      expect(output.string).to match(debugger_output)
     end
     it 'execute swapcase' do
-      repl_output =  /HELLO/
-      repl.handle_input("swapcase('hello')")
-      expect(output.string).to match(repl_output)
+      debugger_output =  /HELLO/
+      debugger.handle_input("swapcase('hello')")
+      expect(output.string).to match(debugger_output)
     end
   end
 
   describe 'whereami' do
     let(:input) do
-      File.expand_path File.join(fixtures_dir, 'sample_start_repl.pp')
+      File.expand_path File.join(fixtures_dir, 'sample_start_debugger.pp')
     end
     let(:options) do
       {
@@ -461,10 +461,10 @@ describe "PuppetRepl" do
     end
 
     it 'runs' do
-      expect(repl.whereami).to match(/\s+5/)
+      expect(debugger.whereami).to match(/\s+5/)
     end
     it 'contains marker' do
-      expect(repl.whereami).to match(/\s+=>\s10/)
+      expect(debugger.whereami).to match(/\s+=>\s10/)
     end
 
   end
@@ -475,15 +475,15 @@ describe "PuppetRepl" do
     end
     if Gem::Version.new(Puppet.version) >= Gem::Version.new('4.0')
       it 'show error message' do
-        repl_output =  /no\ parameter\ named\ 'contact'/
-        repl.handle_input(input)
-        expect(output.string).to match(repl_output)
+        debugger_output =  /no\ parameter\ named\ 'contact'/
+        debugger.handle_input(input)
+        expect(output.string).to match(debugger_output)
       end
     else
       it 'show error message' do
-        repl_output =  /Invalid\ parameter\ contact/
-        repl.handle_input(input)
-        expect(output.string).to match(repl_output)
+        debugger_output =  /Invalid\ parameter\ contact/
+        debugger.handle_input(input)
+        expect(output.string).to match(debugger_output)
       end
     end