+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+## 0.4.0
+  * Rename to puppet-debugger
 ## 0.3.4
-  * Fixes issue with temporay file not being unique
+  * Fixes issue with temporary file not being unique
 ## 0.3.3
   * Adds ability to call breakpoints from within repl
   * Adds newline for each multiline input